Extract parts of connection_switcher to files
